modularize component further. added license
